+++ src/main/java/com/bank/bankaccount/service/AccountsService.java
@@ -3,10 +3,9 @@ package com.bank.bankaccount.service;
 import com.bank.bankaccount.dto.AccountDTO;
 import com.bank.bankaccount.enums.AccountTypes;
 
-import java.math.BigDecimal;
 
 public interface AccountsService {
 
-    AccountDTO createAccount(Long customerId, AccountTypes accountType, BigDecimal initBalance);
+    AccountDTO createAccount(Long customerId, AccountTypes accountType, Double initBalance);
 
 }
